tests/requirements.txt
selenium==4.27.1
webdriver-manager==4.0.2
pytest==8.3.4
pytest-html==4.1.1
pytest-timeout==2.3.1
python-dotenv==1.0.1
requests==2.31.0

